vm: remove all rendering and IO code
The platform-specific code is now in a special z8::player class. This allows
us to load and run carts outside the main application, for instance if we
want to do a text mode version of ZEPTO-8.

Speaking of a text mode version, `zeptool --run cart.p8` will load and run
a cart without opening a new window and output Unicode and ANSI colour codes
to the terminal.