pico8: improve grammar and parser.
We can now count tokens and get almost exactly the same result as PICO-8. The
two differences I noticed are negative numbers that we don’t detect properly,
and a few discrepancies I assume are bugs that I reported upstream.

The grammar has also been improved to reduce backtracking. The correct token
counting is proof of that.