Fix sprite transparency to check for non-zero bits
Pico 8 seems to only care that the transparency nibble in the palette is
non-zero, but before this fix zepto8 was checking that it was exactly 1.
I don't see this difference explicitely spelled out in the offial docs,
but the wiki does say "non-zero".

This seems to only affect carts that are pokeing the draw state. One
example that is fixed is Dank Tomb, where the green background of the
character sprite was visible, but is not after this fix.